backend/requirements.txt
black
pytest
pytest-cov
pylint
python-dateutil
pandas
numpy
fpdf
nbval==0.11.0
ruff
requests
click
databricks
databricks-sql-connector
python-dotenv

flask
sqlalchemy